pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cv-zero-claw-agent"
version = "0.1.0"
description = "Autonomous CV research agent powered by ZeroClaw"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "click>=8.1",
    "python-dotenv>=1.0",
    "feedparser>=6.0",
    "networkx>=3.0",
    "Pillow>=10.0",
    "tiktoken>=0.7.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "markdown>=3.6",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
]

[project.optional-dependencies]
zeroclaw = [
    "zeroclaw-tools>=0.1.0",
]
mlx = [
    "mlx>=0.16.0",
    "mlx-vlm>=0.1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5.0",
    "mypy>=1.10",
]

[project.scripts]
cv-agent = "cv_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cv_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
